fix(entities): migrate primitives to wrapper types for Hibernate correctness
- OrderStateEntity.version: int → Integer (fixes @Version unsaved-value detection)
- OrderItem: int quantity → Integer, double unitPrice → BigDecimal
- Order: double calculateTotalValue() → BigDecimal
- OrderItemDTO: int → Integer, double → BigDecimal
- OrderItemEntity: int → Integer, double → BigDecimal
- OrderPlaced event: double → BigDecimal for all money fields
- BatchJob/BatchJobJpaEntity: int counters → Integer
- ArchUnit test: enforce @Version must use wrapper types
- Updated all 8 test files to use BigDecimal.valueOf() constructors
- Fixed stray '' artifact in pom.xml

Fixes Issue #1 from #139 (JPA @Version optimistic locking)

+++ boilerplate/java/order-service/src/main/java/com/example/orderservice/application/dtos/OrderItemDTO.java
@@ -3,10 +3,11 @@ package com.example.orderservice.application.dtos;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record OrderItemDTO(
     @NotNull UUID productId,
-    @Min(1) int quantity,
-    @Positive double unitPrice
+    @Min(1) Integer quantity,
+    @Positive BigDecimal unitPrice
 ) {}
